Additional changes motivated by reviewer comments.  Rebased against LT master.
