+++ geomesa-core/src/test/scala/geomesa/core/data/FilterToAccumuloTest.scala
@@ -21,8 +21,10 @@ import collection.JavaConversions._
 import com.vividsolutions.jts.geom.{Polygon, Coordinate}
 import geomesa.core.data.FilterToAccumulo._
 import geomesa.core.index.Constants
+import geomesa.utils.geometry.Geometry._
 import geomesa.utils.geotools.Conversions._
 import geomesa.utils.text.WKTUtils
+import geomesa.utils.time.Interval._
 import org.geotools.data.DataUtilities
 import org.geotools.factory.CommonFactoryFinder
 import org.geotools.filter.text.ecql.ECQL
